refactor: finalize code & tests
* Finish fixing tests for IE11
* Move code to the src directory
